Update app version to 1.1 and extract caption display logic into CaptionContentView for improved code organization.
