+++ Livcap/Views/AboutView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct AboutView: View {
     @Environment(\.dismiss) private var dismiss
     
-    private let appVersion = "0.9"
+    private let appVersion = "1.1"
     private let privacyPolicyURL = "https://livcap.app/privacy"
     private let termsURL = "https://livcap.app/terms"
     private let githubURL = "https://github.com/wayne-xyz/Livcap"
