+++ Livcap/Views/CaptionView.swift
@@ -55,9 +55,7 @@ struct CaptionView: View {
             showWindowControls = hovering
         }
         .onAppear {
-            // Reset animation state on view appear
-            resetFirstContentAnimation()
-            
+
             // Test CoreAudioTapEngine when view appears
             if #available(macOS 14.4, *) {
 //                engineExamples.startSystemAudioCapture()
@@ -65,12 +63,7 @@ struct CaptionView: View {
 
             }
         }
-        .onChange(of: captionViewModel.captionHistory.isEmpty) { _, isEmpty in
-            // Reset animation when captions are cleared
-            if isEmpty && captionViewModel.currentTranscription.isEmpty {
-                resetFirstContentAnimation()
-            }
-        }
+
         .onDisappear {
             // Stop audio sources when window closes
             if captionViewModel.isMicrophoneEnabled {
@@ -92,79 +85,13 @@ struct CaptionView: View {
                 .frame(width: 80) // Fixed width for buttons
             
             // Centered content area with auto-scroll
-            ScrollViewReader { proxy in
-                ScrollView {
-                    LazyVStack(alignment: .leading, spacing: 8) {
-                        // Caption history (older sentences at top)
-                        ForEach(captionViewModel.captionHistory.indices, id: \.self) { index in
-                            let entry = captionViewModel.captionHistory[index]
-                            let isFirstContent = index == 0 && captionViewModel.currentTranscription.isEmpty
-                            
-                            Text(entry.text)
-                                .font(.system(size: 22, weight: .medium, design: .rounded))
-                                .foregroundColor(.primary)
-                                .lineSpacing(7)
-                                .padding(.horizontal, 20)
-                                .padding(.vertical, 4)
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                                .background(
-                                    RoundedRectangle(cornerRadius: 8, style: .continuous)
-                                        .fill(Color.clear)
-                                        .opacity(opacityLevel)
-                                )
-                                .offset(y: isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOffset : 0)
-                                .opacity(isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOpacity : 1.0)
-                        }
-                        
-                        // Current transcription (real-time at bottom)
-                        if !captionViewModel.currentTranscription.isEmpty {
-                            let isFirstContent = captionViewModel.captionHistory.isEmpty
-                            
-                            Text(captionViewModel.currentTranscription+"...")
-                                .font(.system(size: 22, weight: .medium, design: .rounded))
-                                .foregroundColor(.primary)
-                                .padding(.horizontal, 20)
-                                .padding(.vertical, 4)
-                                .lineSpacing(7)
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                                .id("currentTranscription")
-                                .offset(y: isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOffset : 0)
-                                .opacity(isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOpacity : 1.0)
-                        }
-                    }
-                    .padding(.horizontal, 16)
-                    .padding(.vertical, 16)
-                }
-                .onChange(of: captionViewModel.currentTranscription) {
-                    // Trigger first content animation for currentTranscription
-                    if !captionViewModel.currentTranscription.isEmpty && !hasShownFirstContentAnimation && captionViewModel.captionHistory.isEmpty {
-                        triggerFirstContentAnimation()
-                    }
-                    
-                    if !captionViewModel.currentTranscription.isEmpty {
-                        withAnimation(.easeOut(duration: 0.3)) {
-                            proxy.scrollTo("currentTranscription", anchor: .bottom)
-                        }
-                    }
-                }
-                .onChange(of: captionViewModel.captionHistory.count) {
-                    // Trigger first content animation for first caption history entry
-                    if !captionViewModel.captionHistory.isEmpty && !hasShownFirstContentAnimation {
-                        triggerFirstContentAnimation()
-                    }
-                    
-                    // Auto-scroll when new caption is added
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                        if let lastEntry = captionViewModel.captionHistory.last {
-                            withAnimation(.easeOut(duration: 0.3)) {
-                                proxy.scrollTo(lastEntry.id, anchor: .bottom)
-                            }
-                        }
-                    }
-                }
-            }
+            CaptionContentView(
+                captionViewModel: captionViewModel,
+                hasShownFirstContentAnimation: $hasShownFirstContentAnimation,
+                firstContentAnimationOffset: $firstContentAnimationOffset,
+                firstContentAnimationOpacity: $firstContentAnimationOpacity
+            )
             .padding(.horizontal, 20) // Equal margins on both sides
-
             
             // Right side buttons: system audio, mic and pin (removed recording button)
             controlButtons()
@@ -191,101 +118,18 @@ struct CaptionView: View {
             .frame(height: 50)
             
             // Main content area with auto-scroll
-            ScrollViewReader { proxy in
-                ScrollView {
-                    LazyVStack(alignment: .leading, spacing: 8) {
-                        // Caption history (older sentences at top)
-                        ForEach(captionViewModel.captionHistory.indices, id: \.self) { index in
-                            let entry = captionViewModel.captionHistory[index]
-                            let isFirstContent = index == 0 && captionViewModel.currentTranscription.isEmpty
-                            
-                            Text(entry.text)
-                                .font(.system(size: 22, weight: .medium, design: .rounded))
-                                .foregroundColor(.primary)
-                                .padding(.horizontal, 20)
-                                .lineSpacing(7)
-                                .padding(.vertical, 4)
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                                .background(
-                                    RoundedRectangle(cornerRadius: 8, style: .continuous)
-                                        .fill(Color.clear)
-                                        .opacity(opacityLevel)
-                                )
-                                .offset(y: isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOffset : 0)
-                                .opacity(isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOpacity : 1.0)
-                        }
-                        
-                        // Current transcription (real-time at bottom)
-                        if !captionViewModel.currentTranscription.isEmpty {
-                            let isFirstContent = captionViewModel.captionHistory.isEmpty
-                            
-                            Text(captionViewModel.currentTranscription+"...")
-                                .font(.system(size: 22, weight: .medium, design: .rounded))
-                                .foregroundColor(.primary)
-                                .padding(.horizontal, 20)
-                                .lineSpacing(7)
-                                .padding(.vertical, 4)
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                                .id("currentTranscription")
-                                .offset(y: isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOffset : 0)
-                                .opacity(isFirstContent && !hasShownFirstContentAnimation ? firstContentAnimationOpacity : 1.0)
-                        }
-                    }
-                    .padding(.horizontal, 16)
-                    .padding(.vertical, 16)
-                }
-                .onChange(of: captionViewModel.currentTranscription) {
-                    // Trigger first content animation for currentTranscription
-                    if !captionViewModel.currentTranscription.isEmpty && !hasShownFirstContentAnimation && captionViewModel.captionHistory.isEmpty {
-                        triggerFirstContentAnimation()
-                    }
-                    
-                    if !captionViewModel.currentTranscription.isEmpty {
-                        withAnimation(.easeOut(duration: 0.3)) {
-                            proxy.scrollTo("currentTranscription", anchor: .bottom)
-                        }
-                    }
-                }
-                .onChange(of: captionViewModel.captionHistory.count) {
-                    // Trigger first content animation for first caption history entry
-                    if !captionViewModel.captionHistory.isEmpty && !hasShownFirstContentAnimation {
-                        triggerFirstContentAnimation()
-                    }
-                    
-                    // Auto-scroll when new caption is added
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                        if let lastEntry = captionViewModel.captionHistory.last {
-                            withAnimation(.easeOut(duration: 0.3)) {
-                                proxy.scrollTo(lastEntry.id, anchor: .bottom)
-                            }
-                        }
-                    }
-                }
-            }
+            CaptionContentView(
+                captionViewModel: captionViewModel,
+                hasShownFirstContentAnimation: $hasShownFirstContentAnimation,
+                firstContentAnimationOffset: $firstContentAnimationOffset,
+                firstContentAnimationOpacity: $firstContentAnimationOpacity
+            )
             .padding(.horizontal, 20) // Equal margins on both sides for centering
             .padding(.bottom, 20)
+        }
+    }
 
-        }
-    }
-    
-    // MARK: - Animation Functions
-    
-    private func triggerFirstContentAnimation() {
-        guard !hasShownFirstContentAnimation else { return }
-        
-        withAnimation(.spring(response: 0.6, dampingFraction: 0.8, blendDuration: 0.1)) {
-            firstContentAnimationOffset = 0
-            firstContentAnimationOpacity = 1.0
-        }
-        
-        hasShownFirstContentAnimation = true
-    }
-    
-    private func resetFirstContentAnimation() {
-        hasShownFirstContentAnimation = false
-        firstContentAnimationOffset = 30
-        firstContentAnimationOpacity = 0
-    }
+
     
     // MARK: - Window Management Functions
